+++ changelog.md
@@ -2,8 +2,9 @@
 #### 4.27.0-SNAPSHOT
   * `ReadOptionsBuilder.addInjectorFactory()` added to allow additional `InjectorFactory's` to be added.
   * `ReadOptionsBuilder.addFieldFilter()` added to allow additional `FieldFilters` to be added.
+  * LRU size control added to `ReadOptionsBuild` and `WriteOptionsBuilder`. These control the LRU size of the cache that maps `Classes` to `Fields`, `Classes` to `Injectors`, and `Classes` to `Accessors.` 
 #### 4.26.0
-  * Performance improvement for `JsonIo`: When using `null` for default `ReadOptions` or `WriteOptions,` the same static instance is used as the `ReadOptions` and `WriteOptions` are immutable.  
+  * Performance improvement for `JsonIo`: When using `null` for default `ReadOptions` or `WriteOptions,` the same static instance is used.  
   * Updated [java-util](https://github.com/jdereg/java-util/blob/master/changelog.md) from `2.10.0` to `2.13.0.`
 #### 4.25.0
   * `JsonParser` now uses an instance-based cache for common values, not a static one.  This will allow for more speed during concurrent parsing.